bot/requirements.txt
attrs==19.1.0
asyncio==3.4.3
aiohttp==3.7.3
proxybroker==0.3.2
aiogram==2.11.2
matplotlib==3.3.2
aiosqlite==0.16.0
tortoise-orm==0.16.17
pydantic==1.7.2